[HER-16] Rename body to row
