+++ src/admin/components/styled-row/styled-row.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { TableRow } from '@material-ui/core'
 import PropTypes from 'prop-types'
 
-import { useStyles } from 'admin/components/row/style'
+import { useStyles } from 'admin/components/styled-row/style'
 
 export const StyledRow = ({ children }) => <TableRow classes={useStyles()}>
   {children}
